refactor upload code by moving it over to new component
