+++ src/components/Customer/PPM/Closeout/WeightTicketUpload/WeightTicketUpload.module.scss
@@ -1,0 +1,38 @@
+@import 'shared/styles/_basics';
+@import 'shared/styles/colors';
+
+
+.WeightTicketUpload {
+  .uploadsTable {
+    @include u-margin-y(2);
+  }
+
+  .uploadTypeHint {
+    @include u-margin-bottom(1);
+  }
+
+  .constructedWeightLink {
+    @include u-margin-top(2);
+    width: fit-content;
+    @include u-text-decoration(no-underline);
+
+    &:hover, &:hover:visited {
+      color: $primary-dark;
+      @include u-text-decoration(no-underline);
+    }
+
+    &:visited {
+      color: $primary;
+    }
+  }
+
+  /* FilePond attempts to do some animating with inline styling but the label never transitions to visible in some
+   browser environments. This might explain past Happo errors where labels were blank in the mobile visual diffs.
+   Maybe we could make this a global fix for orders uploads too. I don't see another way to disable this animation. */
+  :global(.filepond--drop-label) {
+    transform: unset !important;
+    opacity: 100 !important;
+    visibility: visible !important;
+    pointer-events: all !important;
+  }
+}
